VC8用のvcprojにHISpidar4Dが入ってなかったので修正 DRUsb20Sh4.cpp: 型変換warningを対処
git-svn-id: http://springhead.info/spr2/Springhead2/trunk@4609 05cee5c3-a2e9-0310-9523-9dfc2f93dbe1

+++ src/HumanInterface/DRUsb20Sh4.cpp
@@ -93,7 +93,7 @@ void DRUsb20Sh4::UsbUpdate(){
 			}
 		}
 		ct += diff;
-		count[i] = -ct;
+		count[i] = -(long)ct;
 	}
 	//	A/DÇÃì«Ç›èoÇµ
 	for(int i=0; i<8; ++i){
